move cache into different file and add as watch dependency
